separete gui.py into files in ui folder
